refactor(formatter): split Formatter impl into per-rule files under formatter/rules/
Closes #2261
